% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTTPstatus.R
\name{HTTPstatus}
\alias{HTTPstatus}
\title{HTTP status handler}
\usage{
HTTPstatus(status)
}
\arguments{
\item{status}{numeric, the error status}
}
\value{
A list of two elements. `warning` is used to create a
warning message and `rtn` will be used to return to the user.
}
\description{
This function interprets the meaning of the HTTP status and returns
the elements needed to create an appropriate warning and return
something meaningful to the user.
}
\details{
Code 200 should not be sent to this function as this
is only designed to handle the errors (>=400)
}
\examples{
HTTPstatus(400)
}
