% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify}
\alias{identify}
\title{HTTP status handler}
\usage{
identify(key, imageURL, simplify = TRUE)
}
\arguments{
\item{key}{character, your API key (get from https://my.plantnet.org/)}

\item{imageURL}{character, the URL path top the image you want to identify}

\item{simplify}{logical, if `TRUE` the output will be simplified into a
data.frame}
}
\value{
If `simplify` is set to `FALSE` then a list is returned. The
structure of this list can be explored using hte tool at
https://my-api.plantnet.org. If `simplify` is set to `TRUE` (default)
then a data.frame is with three columns for the score, latin name,
and common name. There are as many rows as there are species predictions
from the API.
}
\description{
This function interprets the meaning of the HTTP status and returns
the elements needed to create an appropriate warning and return
something meaningful to the user.
}
\details{
The function uses the PlantNet API. To use this service you need
to have registered an account and generated an API key. All this can be
done here: https://my.plantnet.org/.
}
