% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyPN.R
\name{simplifyPN}
\alias{simplifyPN}
\title{Simplify PlantNet response}
\usage{
simplifyPN(sp)
}
\arguments{
\item{sp}{a single element from the response 'results' section}
}
\value{
A data.frame with three columns for the score, latin name,
and common name. There are as many rows as there are species predictions
from the API
}
\description{
This function is used to simplify the response from the PlantNet
API into a simple table of the important data that a user is
most likely to want.
}
